requirements.txt
google-cloud-speech==2.20.0
google-cloud-dialogflow==2.12.0
google-cloud-texttospeech==2.14.0
python-dotenv==0.21.0
Flask==2.2.2
requests==2.28.1